+++ pytorch_lightning/utilities/__init__.py
@@ -5,7 +5,7 @@ import torch
 
 from pytorch_lightning.utilities.distributed import rank_zero_only, rank_zero_warn, rank_zero_info
 from pytorch_lightning.utilities.apply_func import move_data_to_device
-from pytorch_lightning.utilities.parsing import AttributeDict
+from pytorch_lightning.utilities.parsing import AttributeDict, flatten_dict
 
 try:
     from apex import amp